+++ packages/svelte/src/compiler/phases/2-analyze/validation.js
@@ -476,6 +476,7 @@ export const validation = {
 				grand_parent?.type !== 'SvelteComponent' &&
 				grand_parent?.type !== 'EachBlock' &&
 				grand_parent?.type !== 'AwaitBlock' &&
+				grand_parent?.type !== 'SnippetBlock' &&
 				((grand_parent?.type !== 'RegularElement' && grand_parent?.type !== 'SvelteElement') ||
 					!grand_parent.attributes.some((a) => a.type === 'Attribute' && a.name === 'slot')))
 		) {