+++ packages/svelte/tests/runtime-runes/samples/each-fallback/_config.js
@@ -1,0 +1,17 @@
+import { flushSync } from 'svelte';
+import { test } from '../../test';
+
+export default test({
+	async test({ assert, target, component }) {
+		const [b1] = target.querySelectorAll('button');
+		assert.htmlEqual(target.innerHTML, '<div>abc</div><button>Toggle</button>');
+		flushSync(() => {
+			b1.click();
+		});
+		assert.htmlEqual(target.innerHTML, '<div>Fallback</div><button>Toggle</button>');
+		flushSync(() => {
+			b1.click();
+		});
+		assert.htmlEqual(target.innerHTML, '<div>abc</div><button>Toggle</button>');
+	}
+});