+++ packages/svelte/tests/runtime-runes/samples/props-alias/Counter.svelte
@@ -1,0 +1,5 @@
+<script>
+	let { count: definedCount } = $props();
+</script>
+
+<button on:click={() => definedCount++}>{definedCount}</button>